[REFACTOR] DatabaseGatewayのパッケージ変更
