+++ src/main/java/com/github/unchama/seichiassist/commands/AchieveCommand.java
@@ -4,7 +4,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.UUID;
 
-import com.github.unchama.seichiassist.DatabaseGateway;
+import com.github.unchama.seichiassist.database.DatabaseGateway;
 import net.md_5.bungee.api.ChatColor;
 
 import org.bukkit.Bukkit;
