+++ src/main/java/com/github/unchama/seichiassist/commands/lastquitCommand.java
@@ -1,7 +1,7 @@
 package com.github.unchama.seichiassist.commands;
 
 import com.github.unchama.seichiassist.SeichiAssist;
-import com.github.unchama.seichiassist.DatabaseGateway;
+import com.github.unchama.seichiassist.database.DatabaseGateway;
 import com.github.unchama.seichiassist.util.Util;
 import net.md_5.bungee.api.ChatColor;
