+++ src/main/java/com/github/unchama/seichiassist/listener/PlayerJoinListener.java
@@ -17,7 +17,7 @@ import org.bukkit.inventory.ItemStack;
 
 import com.github.unchama.seichiassist.ActiveSkill;
 import com.github.unchama.seichiassist.SeichiAssist;
-import com.github.unchama.seichiassist.DatabaseGateway;
+import com.github.unchama.seichiassist.database.DatabaseGateway;
 import com.github.unchama.seichiassist.data.PlayerData;
 import com.github.unchama.seichiassist.util.Util;
